+++ src/components/Select/index.ts
@@ -1,3 +1,3 @@
-// Select from copy-ui @ 2025-10-31.
+// Select from copy-ui @ 2025-11-04.
 
 export { Select } from "./select";